Use user ID to log in, adjustments to messages, ASCII art
The program now uses the student's User ID (format: name.surname), and retrieves the name from the string to greet the student. Additionally, the "You are due X free coffee(s)" message has been tweaked so that it displays "coffee" when 1 coffee is due, "coffees" when 2+ coffees are due. Lastly, ASCII art has been used to indicate that the coffee has been served.