First version
This program is a booking system for a hotel. Customers can define book-in and book-out dates, and the code will look if 1 of the 4 rooms is available for booking. However, the code is rather simple and will only check if the new booking can fit after the last booking of all the rooms (i.e. if booking is for a time frame between two pre-existing bookings, it will say the room is not available for booking). Additionally, the program also allows the managment to easily modify the per-night pricing.